requirement.txt
yfinance
pandas
numpy
scikit-learn
matplotlib
pyyaml
arch
xgboost
